+++ CHANGELOG.md
@@ -2,5 +2,15 @@
 All notable changes to this project will be documented in
 this file.
 
+## 1.1 - 2022/02/04 23:41 GMT+08:00
+### Changed
+* Changed the Output Folder textbox to a Directory textbox
+to allow the user to browse to a folder instead of manually
+typing the path (thanks to JEJ)
+### Fixed
+* Fixed bug where if the specified path doesn't work, the
+script fails to create or modify the CSV file in the scripts
+folder
+
 ## 1.0 - 2022/01/21 05:39 GMT+08:00
 * initial release